backend/pyproject.toml
[project]
name = "helm-backend"
version = "0.1.0"
description = "Personal Academic Planner + Budgeting API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0", # Sync driver for Alembic migrations
    "alembic>=1.14.0",
    # Data Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "google-auth>=2.37.0",
    "requests>=2.32.0", # Required by google-auth transport
    # HTTP client (for Google token verification)
    "httpx>=0.28.0",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
